backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic==1.13.0",
    "asyncpg==0.29.0",
    "bcrypt==4.1.2",
    "fastapi==0.104.1",
    "httpx>=0.26.0,<0.27.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.5.2",
    "pydantic-settings==2.1.0",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=7.0.0",
    "pytest-httpx==0.28.0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "sqlalchemy[asyncio]==2.0.23",
    "structlog==23.2.0",
    "tenacity==8.2.3",
    "uvicorn[standard]==0.24.0",
]
